services/trading-engine/app/routers/auth.py
"""
Authentication endpoints
"""
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
from sqlalchemy.ext.asyncio import AsyncSession
from jose import JWTError, jwt
from passlib.context import CryptContext
from datetime import datetime, timedelta
import logging
from typing import Optional, Dict, Any

from ..database import get_db
from ..config import settings

router = APIRouter()
logger = logging.getLogger(__name__)

# Password hashing
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth2 scheme
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/api/v1/auth/token")


def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    """
    Create JWT access token
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(hours=settings.JWT_EXPIRATION_HOURS)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)
    return encoded_jwt


@router.post("/register")
async def register(
    email: str,
    username: str,
    password: str,
    db: AsyncSession = Depends(get_db)
) -> Dict[str, Any]:
    """
    Register a new user
    """
    # TODO: Implement user registration
    return {
        "message": "User registration endpoint - to be implemented",
        "email": email,
        "username": username
    }


@router.post("/token")
async def login(
    form_data: OAuth2PasswordRequestForm = Depends(),
    db: AsyncSession = Depends(get_db)
) -> Dict[str, str]:
    """
    Login endpoint that returns JWT token
    """
    # TODO: Implement user authentication
    # For now, return a dummy token
    access_token = create_access_token(
        data={"sub": form_data.username}
    )
    return {
        "access_token": access_token,
        "token_type": "bearer"
    }


@router.get("/me")
async def get_current_user(token: str = Depends(oauth2_scheme)) -> Dict[str, Any]:
    """
    Get current user information
    """
    # TODO: Implement token validation and user retrieval
    return {
        "message": "Current user endpoint - to be implemented",
        "token": token[:20] + "..."
    }